Should use the LineNumber type

+++ sources/linenumber.hpp
@@ -3,6 +3,5 @@
 #include <cstdint>
 
 namespace BigWord {
-// The line number nested type name.
-using LineNumber = int64_t;
+using LineNumber = uint32_t;  // The line number nested type name.
 }  // namespace BigWord
